requirements.txt
Django==2.2.6
django-braces==1.13.0
django-crispy-forms==1.7.2
-e git+https://github.com/novapost/django-highcharts.git@34dd1b644325aaf8d9772feaf09827c730e8b154#egg=django_highcharts
django-postgres-copy==2.4.2
django-redis==4.10.0
djangorestframework==3.10.3
psycopg2==2.8.3
pytz==2019.2
redis==3.3.8
six==1.12.0
sqlparse==0.3.0
